Threshold is not a part of analytic unit #619 (#620)
