+++ server/src/models/detection_model.ts
@@ -1,4 +1,4 @@
-import { AnalyticUnitId } from './analytic_unit_model';
+import { AnalyticUnitId } from './analytic_units';
 import { Collection, makeDBQ } from '../services/data_service';
 
 import * as _ from 'lodash';
