+++ server/src/services/data_puller.ts
@@ -1,5 +1,5 @@
 import { AnalyticsTask, AnalyticsTaskType } from '../models/analytics_task_model';
-import * as AnalyticUnit from '../models/analytic_unit_model';
+import * as AnalyticUnit from '../models/analytic_units';
 import * as AnalyticUnitCache from '../models/analytic_unit_cache_model';
 import { AnalyticsService } from './analytics_service';
 import { HASTIC_API_KEY } from '../config';
